Deprecate ::MichaelScott
