+++ doc/unreleased/tv_shows/michael_scott.md
@@ -1,0 +1,7 @@
+# Faker::TvShows::MichaelScott
+
+Available since version 1.9.0.
+
+```ruby
+Faker::TvShows::MichaelScott.quote #=> "I am Beyoncé, always."
+```
