+++ lib/faker/deprecate/michael_scott.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Faker
+  class MichaelScott < Base
+    class << self
+      extend Gem::Deprecate
+
+      def quote
+        Faker::TvShows::MichaelScott.quote
+      end
+
+      deprecate :quote, 'Faker::TvShows::MichaelScott.quote', 2018, 12
+    end
+  end
+end
